Cleaned MSWC dataset files
